Add 'center' keyword; allow 'sample', 'precise', 'center' and 'globallycoherent' as identifiers (#1406)
* Add 'center' keyword as an interpolation modifier

FXC allows the interpolation modifier 'center' (equivalent to 'linear') and DXC should too. At places where interpolation modifier is not expected 'center' can still be used as an identifier. This change makes it easier to switch from FXC to DXC. Add warnings when center is used with centroid or sample. Update varmod-syntax.hlsl test to include 'center'.

* Allow 'sample', 'precise' and 'globallycoherent' as identifiers

In DXC 'sample', 'precise' and 'globallycoherent' were treated as a keywords and could not be used as identifiers. FXC allowed this.
This change makes it easier to switch from FXC to DXC.

* Update fxc errors in varmods-syntax.hlsl based on fxc version 10.0.17713.