[linux-port] gtest interface for HLSL tests (#1413)
On Windows, HLSL tests run using TAEF. Since that's not available
on other platforms, this change adapts the tests to work on gtest
framework with minimal alterations to the tests themselves.

TestMain.cpp is pilfered and enhanced from Spirv tests. It provides
a starting point for the gtest binary to initialize variables and
launch tests, though gtest hides most of that.

Derived from SpirvTestOptions, this lists a number of potential
options from the command line and provides a crude, but effective
mechanism to access the as needed.

With this commit, you can run clang-hlsl-tests, but it does't really
do anything. Tests will follow shortly.