[spirv] Support space only :register() (#1503)
This commit relaxes the syntax of :register() to omit the type and
number and only keep space. That is, aside from :register(xX, spaceY),
we now also support :register(spaceY). This is useful for auto-assign
binding numbers with explicit set numbers for Vulkan.

For DXIL side, an error will be emitted currently if using space-only
register.