[DxilPIXPasses] Assigns unique ID to each instruction. (#1888)
This PR modifies the DxilAnnotateWithVirtualRegister pass to also annotate
each instruction in the DXIL module with a unique ID. This ID can then be
used by trace debuggers to uniquely identify an instruction in DXIL module.

This PR also modifies the DxilDebugInstrumentation pass to output this
unique ID on each entry in the debug trace. The debug trace is also expanded
to include the Value Ordinal (i.e., the virtual register modified by the
instruction).

Finally, this PR bundles a small bugfix to the dxassembler. The bug prevented
the DxcAssembler to add the debug DXIL to the container it assembles.