Added an error for constant array indices being out of bounds. (#1931)
Out-of-bound array indices cause SROA to generate undef values. If returning that value, the finalize dxil pass will get rid of the storeOutput. See issue #1879

We do produce a warning on out-of-bounds errors, but FXC makes it an error and that's probably sensible due to the cascading effects of such undefs. We can only do that on constant expressions though, so the undef scenario can still happen for constant-valued indices resulting from later IR passes.