Revert lying about resource sizes in debug info. (#2114)
Pull request #2086 lies about the size of resources to ensure that backends can always expect 32 bits in the debug info. However, this causes warnings/errors in code containing structs of a resource type and a numerical type, because SROA uses the LLVM size rather than the clang size, and those are expected to be in sync. So the dwarf is going to say that the resource is 32 bits, but LLVM will think it is 64 bits (given our current definition of the struct).