PIX shader access tracking pass: instrument all functions in module (#2148)
The actual problem here is that the instrumentation was only adding a CreateHandle call for the entrypoint function, but emitting UAV writes for all functions. This didn't work for HS shaders (that can have two entry points).
Now, a distinct CreateHandle with the same parameters is emitted for each function in the module.
This pass won't be called on libraries (yet), and will probably need updating anyway when it does.